Added user_roles.py to top level directory
