--- api/user_roles/admin_roles.py
@@ -1,5 +1,0 @@
-admin_roles = ["admin"]
-
-
-def is_admin(user_role):
-    return user_role in admin_roles
